refactor(instance): rename instances to instance
