+++ pkg/cmd/instance/add.go
@@ -1,4 +1,4 @@
-package instances
+package instance
 
 import (
 	"fmt"
@@ -29,7 +29,7 @@ func NewAddCmd(f *cmdutil.Factory, runF func(*AddOptions) error) *cobra.Command 
 		Use: "add",
 		Args: func(cmd *cobra.Command, args []string) error {
 			if err := cobra.MinimumNArgs(1)(cmd, args); err != nil {
-				return fmt.Errorf("ERROR: specify an API hostname\n$ inst instances add API_HOSTNAME")
+				return fmt.Errorf("ERROR: specify an API hostname\n$ inst instance add API_HOSTNAME")
 			}
 			if err := instance.HostnameValidator(args[0]); err != nil {
 				return fmt.Errorf("error parsing API hostname %w", err)
@@ -42,10 +42,10 @@ func NewAddCmd(f *cmdutil.Factory, runF func(*AddOptions) error) *cobra.Command 
 		`),
 		Example: heredoc.Doc(`
 			# add a local instance as the default one
-			$ inst instances add instill.localhost --default
+			$ inst instance add instill.localhost --default
 
 			# add a cloud instance
-			$ inst instances add api.instill.tech \
+			$ inst instance add api.instill.tech \
 				--oauth2 auth.instill.tech \
 				--audience https://instill.tech \
 				--issuer https://auth.instill.tech/ \
