+++ pkg/cmd/instance/set_default.go
@@ -1,4 +1,4 @@
-package instances
+package instance
 
 import (
 	"fmt"
@@ -12,7 +12,7 @@ import (
 	"github.com/instill-ai/cli/pkg/iostreams"
 )
 
-type RemoveOptions struct {
+type SetDefaultOptions struct {
 	IO             *iostreams.IOStreams
 	Config         config.Config
 	MainExecutable string
@@ -20,29 +20,30 @@ type RemoveOptions struct {
 	APIHostname    string
 }
 
-func NewRemoveCmd(f *cmdutil.Factory, runF func(*RemoveOptions) error) *cobra.Command {
-	opts := &RemoveOptions{
+func NewSetDefaultCmd(f *cmdutil.Factory, runF func(*SetDefaultOptions) error) *cobra.Command {
+	opts := &SetDefaultOptions{
 		IO: f.IOStreams,
 	}
 
 	cmd := &cobra.Command{
-		Use: "remove",
+		Use: "set-default",
 		Args: func(cmd *cobra.Command, args []string) error {
 			if err := cobra.MinimumNArgs(1)(cmd, args); err != nil {
-				return fmt.Errorf("ERROR: specify an API hostname\n$ inst instances remove API_HOSTNAME")
+				return fmt.Errorf("ERROR: specify an API hostname\n$ inst instance set-default API_HOSTNAME")
 			}
 			if err := instance.HostnameValidator(args[0]); err != nil {
 				return fmt.Errorf("error parsing API hostname %w", err)
 			}
 			return nil
 		},
-		Short: "Remove an existing instance",
+		Short: "Mark an instance as the default one",
 		Long: heredoc.Docf(`
-			Remove an existing Instill AI instance, either Cloud or Core.
+			Mark an instance as the default one for commands like "auth" and "api".
+			Can be optionally overloaded with the --hostname param.
 		`),
 		Example: heredoc.Doc(`
-			# remove the local instance instance
-			$ inst instances remove instill.localhost
+			# make instill.localhost the default instance
+			$ inst instance set-default instill.localhost
 		`),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			cfg, err := f.Config()
@@ -55,25 +56,25 @@ func NewRemoveCmd(f *cmdutil.Factory, runF func(*RemoveOptions) error) *cobra.Co
 				opts.Interactive = true
 			}
 			opts.APIHostname = args[0]
+
 			opts.MainExecutable = f.Executable()
 			if runF != nil {
 				return runF(opts)
 			}
 
-			return RunRemove(opts)
+			return runSetDefault(opts)
 		},
 	}
 
 	return cmd
 }
 
-func RunRemove(opts *RemoveOptions) error {
+func runSetDefault(opts *SetDefaultOptions) error {
 	hosts, err := opts.Config.HostsTyped()
 	if err != nil {
 		return err
 	}
 
-	// TODO use cfg.HostnameExists()
 	apiHost := opts.APIHostname
 	var host *config.HostConfigTyped
 	for _, h := range hosts {
@@ -86,13 +87,14 @@ func RunRemove(opts *RemoveOptions) error {
 		return fmt.Errorf("ERROR: instance '%s' does not exists", apiHost)
 	}
 
-	opts.Config.UnsetHost(opts.APIHostname)
-	err = opts.Config.Write()
+	host.IsDefault = true
+
+	err = opts.Config.SaveTyped(host)
 	if err != nil {
-		return fmt.Errorf("error removing hostname '%s' - %w", opts.APIHostname, err)
+		return fmt.Errorf("ERROR: failed to set instance '%s' as the default one:\n%w", opts.APIHostname, err)
 	}
 
-	cmdutil.P(opts.IO, "Instance '%s' has been removed\n", opts.APIHostname)
+	cmdutil.P(opts.IO, "Instance '%s' has been set as the default one\n", host.APIHostname)
 
 	return nil
 }
